+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.4] - 2026-02-11
+
 ### Fixed
 - Cover state stuck on "opening"/"closing" after reaching fully open/closed position
 - Cover state stuck on "opening"/"closing" after stopping at intermediate position (via stop command or external stop)